add shader debug board, into source mesh
The asowem debug shader from: https://gist.github.com/aras-p/657a4947bb8a4549fda53a84bb91fb25

+++ OF test0/Assets/Projects/MUtility/VoxelGenerationUtility.cs
@@ -144,7 +144,7 @@ namespace MUtility
          *  |     |
          *  0-----2
          */
-        public static readonly int2[] uv_4p =
+        public static readonly float2[] uv_4p =
         {
             new int2( 0, 0 ), //00
             new int2( 0, 1 ), //01
